index.js: will be used for website location/redirection since it provides a smoother transition.
login.php: Added and will check if the user exists within the database along with their hashed password.

signup.pfp: Added and will check if the user exists before adding them to the database. Will hash their password before storing it in the database

signup.js: Implemented the regex for name and password in a separate file for readability.

config.pfp: Automatically connects us to the database and will be included in signup and login html files.